+++ app/src/main/java/org/simple/clinic/di/PagingSize.kt
@@ -1,0 +1,12 @@
+package org.simple.clinic.di
+
+import javax.inject.Qualifier
+
+@Qualifier
+annotation class PagingSize(val value: Page) {
+
+  enum class Page {
+    AllRecentPatients
+  }
+}
+