+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Explicitly pin OkHttp version to 4.X
 - Use macos version 11.0 to run instrumentation tests on CI
 - Migrate `RecentPatientsScreen` to a fragment
+- Paginate recent patients list
 
 ### Changes
 - Updated translations: `pa-IN`, `hi-IN`, `te-IN`, `kn-IN`, `mr-IN`, `pa-IN`, `te-IN`, `sid-ET`, `kn-IN`, `ta-IN`, `bn-BD`